refactor: 타입 구조 개선 및 Mapper 추가
- entities/location/types/api.ts 생성 (도메인 타입 재사용)
- API 변환 로직을 locationMapper로 분리
- 중복된 Response 타입 제거
- fetchRecommendationId, fetchRecommendationResult에서 새 타입 사용
